// Frontend/src/app/api/api-reference.ts
export class Response implements IResponse {
    id?: string;
    title?: string;
    notes?: GetNotesModel[];
    remainder?: Date | undefined;
    isRemind?: boolean;
    position!: number;
    owner?: string;
    shareTime?: Date | undefined;

    constructor(data?: IResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.title = _data["title"];
            if (Array.isArray(_data["notes"])) {
                this.notes = [] as any;
                for (let item of _data["notes"])
                    this.notes!.push(GetNotesModel.fromJS(item));
            }
            this.remainder = _data["remainder"] ? new Date(_data["remainder"].toString()) : <any>undefined;
            this.isRemind = _data["isRemind"];
            this.position = _data["position"];
            this.owner = _data["owner"];
            this.shareTime = _data["shareTime"] ? new Date(_data["shareTime"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): Response {
        data = typeof data === 'object' ? data : {};
        let result = new Response();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        if (Array.isArray(this.notes)) {
            data["notes"] = [];
            for (let item of this.notes)
                data["notes"].push(item.toJSON());
        }
        data["remainder"] = this.remainder ? this.remainder.toISOString() : <any>undefined;
        data["isRemind"] = this.isRemind;
        data["position"] = this.position;
        data["owner"] = this.owner;
        data["shareTime"] = this.shareTime ? this.shareTime.toISOString() : <any>undefined;
        return data;
    }
}

export interface IResponse {
    id?: string;
    title?: string;
    notes?: GetNotesModel[];
    remainder?: Date | undefined;
    isRemind?: boolean;
    position: number;
    owner?: string;
    shareTime?: Date | undefined;
}

export class GetNotesModel implements IGetNotesModel {
    id?: string;
    content?: string;
    isComplete!: boolean;
    toDoListId?: string;
    position!: number;

    constructor(data?: IGetNotesModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.content = _data["content"];
            this.isComplete = _data["isComplete"];
            this.toDoListId = _data["toDoListId"];
            this.position = _data["position"];
        }
    }

    static fromJS(data: any): GetNotesModel {
        data = typeof data === 'object' ? data : {};
        let result = new GetNotesModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["content"] = this.content;
        data["isComplete"] = this.isComplete;
        data["toDoListId"] = this.toDoListId;
        data["position"] = this.position;
        return data;
    }
}

export interface IGetNotesModel {
    id?: string;
    content?: string;
    isComplete: boolean;
    toDoListId?: string;
    position: number;
}